Fixes based on conversation
